+++ OpenSvg.Gtfs/GtfsShapePoint.cs
@@ -1,7 +1,7 @@
 ﻿using OpenSvg.GeoJson;
 
 namespace OpenSvg.Gtfs;
-public record struct GtfsShapePoint(string ID, Coordinate Coordinate, int Sequence = 0, double Distance = 0)
+public record struct GtfsShapePoint(string ID, Coordinate Coordinate, int Sequence = 0, float Distance = 0)
 {
